+++ packages/blazor-workspace/NimbleBlazor/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Wed, 11 Sep 2024 21:31:24 GMT and should not be manually modified.
+This log was last generated on Wed, 18 Sep 2024 20:24:09 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 18.2.5
+
+Wed, 18 Sep 2024 20:24:09 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v32.2.6
 
 ## 18.2.4